Multiple correlation test

Calculates correlations between multiple variables.

Usage

mcor_test(
  x,
  y = NULL,
  estimate = TRUE,
  p.value = FALSE,
  method = "spearman",
  method_adjust = "BH"
)

Arguments

x

Data frame containing numerical variables.

y

Data frame containing numerical variables. If NULL, correlations are calculated within x.

estimate

Logical specifying whether to return correlation coefficients.

p.value

Logical specifying whether to return adjusted p-values.

method

Character specifying the correlation method: pearson, kendall, or spearman.

method_adjust

Character specifying the p-value adjustment method.

Value

Depending on the values of estimate and p.value, one of the following:

estimate = TRUE, p.value = FALSE

A numeric matrix of correlation coefficients, with columns corresponding to variables in x and rows to variables in y.

estimate = FALSE, p.value = TRUE

A numeric matrix of adjusted p-values, with columns corresponding to variables in x and rows to variables in y.

estimate = TRUE, p.value = TRUE

A named list with two elements:

estimate

Numeric matrix of correlation coefficients.

p.value

Numeric matrix of adjusted p-values.

Examples

library(magrittr)
x0 <- runif(20)
x <- lapply(
    c(1, -1),
    function(i) sapply(seq(10), function(j) x0 * i + runif(10, max = 1))
) %>%
    Reduce(cbind, .) %>%
    set_colnames(paste("Variable", seq(20)))
y <- lapply(
    c(1, -1),
    function(i) sapply(seq(10), function(j) x0 * i + runif(10, max = 1))
) %>%
    Reduce(cbind, .) %>%
    set_colnames(paste("Variable", seq(20))) %>%
    .[, seq(5)]
mcor_test(x)
#>             Variable 1  Variable 2 Variable 3 Variable 4  Variable 5
#> Variable 1   1.0000000  0.36992481  0.2586466  0.4150376  0.38045113
#> Variable 2   0.3699248  1.00000000  0.1759398  0.1984962 -0.05263158
#> Variable 3   0.2586466  0.17593985  1.0000000  0.9248120  0.66165414
#> Variable 4   0.4150376  0.19849624  0.9248120  1.0000000  0.73533835
#> Variable 5   0.3804511 -0.05263158  0.6616541  0.7353383  1.00000000
#> Variable 6   0.4090226  0.28721805  0.2646617  0.3669173  0.30075188
#> Variable 7   0.6571429  0.38045113  0.5142857  0.6917293  0.46917293
#> Variable 8   0.5203008  0.31729323  0.5338346  0.5278195  0.43007519
#> Variable 9   0.4932331  0.10977444  0.8240602  0.8150376  0.73082707
#> Variable 10  0.2225564  0.67218045  0.2721805  0.3187970  0.04210526
#> Variable 11 -0.4646617 -0.01954887 -0.6060150 -0.5834586 -0.51428571
#> Variable 12 -0.5413534 -0.01654135 -0.4180451 -0.6165414 -0.55187970
#> Variable 13 -0.5834586 -0.25563910 -0.6691729 -0.7473684 -0.37894737
#> Variable 14 -0.3248120 -0.35488722 -0.7458647 -0.6195489 -0.52180451
#> Variable 15 -0.2601504 -0.03157895 -0.4120301 -0.4436090 -0.45263158
#> Variable 16 -0.8150376 -0.18646617 -0.3789474 -0.5157895 -0.55939850
#> Variable 17 -0.4887218 -0.07368421 -0.5503759 -0.6977444 -0.71879699
#> Variable 18 -0.6105263 -0.23007519 -0.7714286 -0.8060150 -0.72180451
#> Variable 19 -0.3323308 -0.10676692 -0.6195489 -0.5458647 -0.50075188
#> Variable 20 -0.5458647 -0.10375940 -0.6015038 -0.6105263 -0.56090226
#>              Variable 6 Variable 7  Variable 8 Variable 9 Variable 10
#> Variable 1   0.40902256  0.6571429  0.52030075  0.4932331  0.22255639
#> Variable 2   0.28721805  0.3804511  0.31729323  0.1097744  0.67218045
#> Variable 3   0.26466165  0.5142857  0.53383459  0.8240602  0.27218045
#> Variable 4   0.36691729  0.6917293  0.52781955  0.8150376  0.31879699
#> Variable 5   0.30075188  0.4691729  0.43007519  0.7308271  0.04210526
#> Variable 6   1.00000000  0.7834586  0.05864662  0.4075188  0.65263158
#> Variable 7   0.78345865  1.0000000  0.33233083  0.6586466  0.49473684
#> Variable 8   0.05864662  0.3323308  1.00000000  0.6150376 -0.01954887
#> Variable 9   0.40751880  0.6586466  0.61503759  1.0000000  0.19398496
#> Variable 10  0.65263158  0.4947368 -0.01954887  0.1939850  1.00000000
#> Variable 11 -0.71578947 -0.6526316 -0.13233083 -0.5729323 -0.39097744
#> Variable 12 -0.47669173 -0.6390977 -0.51428571 -0.6045113 -0.27067669
#> Variable 13 -0.69624060 -0.8451128 -0.40150376 -0.7368421 -0.49774436
#> Variable 14 -0.50977444 -0.5458647 -0.28120301 -0.6766917 -0.51278195
#> Variable 15 -0.36240602 -0.2691729 -0.65864662 -0.3789474 -0.10075188
#> Variable 16 -0.30977444 -0.5022556 -0.76691729 -0.5954887 -0.12932331
#> Variable 17 -0.52180451 -0.5112782 -0.47368421 -0.5759398 -0.38345865
#> Variable 18 -0.61203008 -0.8315789 -0.41353383 -0.9037594 -0.37593985
#> Variable 19 -0.20300752 -0.4721805 -0.65864662 -0.5954887  0.09323308
#> Variable 20 -0.25864662 -0.4000000 -0.73984962 -0.6631579 -0.18796992
#>             Variable 11 Variable 12 Variable 13 Variable 14 Variable 15
#> Variable 1  -0.46466165 -0.54135338  -0.5834586  -0.3248120 -0.26015038
#> Variable 2  -0.01954887 -0.01654135  -0.2556391  -0.3548872 -0.03157895
#> Variable 3  -0.60601504 -0.41804511  -0.6691729  -0.7458647 -0.41203008
#> Variable 4  -0.58345865 -0.61654135  -0.7473684  -0.6195489 -0.44360902
#> Variable 5  -0.51428571 -0.55187970  -0.3789474  -0.5218045 -0.45263158
#> Variable 6  -0.71578947 -0.47669173  -0.6962406  -0.5097744 -0.36240602
#> Variable 7  -0.65263158 -0.63909774  -0.8451128  -0.5458647 -0.26917293
#> Variable 8  -0.13233083 -0.51428571  -0.4015038  -0.2812030 -0.65864662
#> Variable 9  -0.57293233 -0.60451128  -0.7368421  -0.6766917 -0.37894737
#> Variable 10 -0.39097744 -0.27067669  -0.4977444  -0.5127820 -0.10075188
#> Variable 11  1.00000000  0.36842105   0.7127820   0.7533835  0.34736842
#> Variable 12  0.36842105  1.00000000   0.6932331   0.2225564  0.53684211
#> Variable 13  0.71278195  0.69323308   1.0000000   0.5548872  0.39097744
#> Variable 14  0.75338346  0.22255639   0.5548872   1.0000000  0.15488722
#> Variable 15  0.34736842  0.53684211   0.3909774   0.1548872  1.00000000
#> Variable 16  0.35939850  0.78796992   0.5593985   0.2436090  0.61954887
#> Variable 17  0.54586466  0.74586466   0.5879699   0.3503759  0.75338346
#> Variable 18  0.76992481  0.55789474   0.7939850   0.7924812  0.24060150
#> Variable 19  0.45413534  0.37744361   0.4135338   0.5954887  0.41052632
#> Variable 20  0.43609023  0.78345865   0.6090226   0.4466165  0.62406015
#>             Variable 16 Variable 17 Variable 18 Variable 19 Variable 20
#> Variable 1   -0.8150376 -0.48872180  -0.6105263 -0.33233083  -0.5458647
#> Variable 2   -0.1864662 -0.07368421  -0.2300752 -0.10676692  -0.1037594
#> Variable 3   -0.3789474 -0.55037594  -0.7714286 -0.61954887  -0.6015038
#> Variable 4   -0.5157895 -0.69774436  -0.8060150 -0.54586466  -0.6105263
#> Variable 5   -0.5593985 -0.71879699  -0.7218045 -0.50075188  -0.5609023
#> Variable 6   -0.3097744 -0.52180451  -0.6120301 -0.20300752  -0.2586466
#> Variable 7   -0.5022556 -0.51127820  -0.8315789 -0.47218045  -0.4000000
#> Variable 8   -0.7669173 -0.47368421  -0.4135338 -0.65864662  -0.7398496
#> Variable 9   -0.5954887 -0.57593985  -0.9037594 -0.59548872  -0.6631579
#> Variable 10  -0.1293233 -0.38345865  -0.3759398  0.09323308  -0.1879699
#> Variable 11   0.3593985  0.54586466   0.7699248  0.45413534   0.4360902
#> Variable 12   0.7879699  0.74586466   0.5578947  0.37744361   0.7834586
#> Variable 13   0.5593985  0.58796992   0.7939850  0.41353383   0.6090226
#> Variable 14   0.2436090  0.35037594   0.7924812  0.59548872   0.4466165
#> Variable 15   0.6195489  0.75338346   0.2406015  0.41052632   0.6240602
#> Variable 16   1.0000000  0.74285714   0.5338346  0.43007519   0.8631579
#> Variable 17   0.7428571  1.00000000   0.5609023  0.21654135   0.7007519
#> Variable 18   0.5338346  0.56090226   1.0000000  0.57142857   0.5563910
#> Variable 19   0.4300752  0.21654135   0.5714286  1.00000000   0.5789474
#> Variable 20   0.8631579  0.70075188   0.5563910  0.57894737   1.0000000
mcor_test(
    x,
    y,
    p.value = TRUE,
    method = "pearson",
    method_adjust = "bonferroni"
)
#> $estimate
#>            Variable 1  Variable 2 Variable 3 Variable 4 Variable 5 Variable 6
#> Variable 1  0.4780656 -0.03132622  0.1025086  0.2975384  0.5506476  0.4724215
#> Variable 2  0.5050967  0.23569605  0.4219412  0.6779986  0.5922804  0.5384219
#> Variable 3  0.3732670  0.08866466  0.7388875  0.6586713  0.6838093  0.5587073
#> Variable 4  0.5972672  0.13046332  0.7693653  0.7194044  0.6266090  0.3478007
#> Variable 5  0.4381090  0.48950445  0.5960861  0.6927373  0.5606594  0.2062807
#>            Variable 7 Variable 8 Variable 9 Variable 10 Variable 11 Variable 12
#> Variable 1  0.4339948  0.3396870  0.1779277   0.1030724  -0.5065104  -0.5748629
#> Variable 2  0.5400468  0.5664006  0.4089203   0.4349587  -0.4392688  -0.8738868
#> Variable 3  0.4454466  0.4896835  0.8595350   0.2226704  -0.7475199  -0.3645730
#> Variable 4  0.6407058  0.4692962  0.8614827   0.1452485  -0.6592442  -0.5260074
#> Variable 5  0.6722295  0.7188243  0.5570295   0.2200366  -0.2989074  -0.3706614
#>            Variable 13 Variable 14 Variable 15 Variable 16 Variable 17
#> Variable 1  -0.2127705  -0.3142392  -0.6824620  -0.6076908  -0.5450868
#> Variable 2  -0.5893198  -0.1902300  -0.7548822  -0.7424850  -0.8567746
#> Variable 3  -0.6514698  -0.7101213  -0.4860625  -0.4882720  -0.6783940
#> Variable 4  -0.7381429  -0.8094427  -0.1040772  -0.5444592  -0.4189003
#> Variable 5  -0.3804936  -0.4359480  -0.4944950  -0.5169442  -0.4116178
#>            Variable 18 Variable 19 Variable 20
#> Variable 1  -0.3885464  -0.4366791  -0.4541989
#> Variable 2  -0.4144272  -0.2562594  -0.7136643
#> Variable 3  -0.7614733  -0.4397951  -0.6057193
#> Variable 4  -0.8708029  -0.7280926  -0.6695048
#> Variable 5  -0.5639371  -0.5958650  -0.2908196
#> 
#> $p.value
#>            Variable 1 Variable 2  Variable 3 Variable 4 Variable 5 Variable 6
#> Variable 1  1.0000000          1 1.000000000 1.00000000 1.00000000          1
#> Variable 2  1.0000000          1 1.000000000 0.10188266 0.59315848          1
#> Variable 3  1.0000000          1 0.019814035 0.15884577 0.08858755          1
#> Variable 4  0.5425049          1 0.007325976 0.03499237 0.31131895          1
#> Variable 5  1.0000000          1 0.554167675 0.07102686 1.00000000          1
#>            Variable 7 Variable 8   Variable 9 Variable 10 Variable 11
#> Variable 1  1.0000000 1.00000000 1.0000000000           1   1.0000000
#> Variable 2  1.0000000 0.92241157 1.0000000000           1   1.0000000
#> Variable 3  1.0000000 1.00000000 0.0001200519           1   0.0151586
#> Variable 4  0.2337190 1.00000000 0.0001066659           1   0.1568377
#> Variable 5  0.1167095 0.03556477 1.0000000000           1   1.0000000
#>             Variable 12 Variable 13 Variable 14 Variable 15 Variable 16
#> Variable 1 8.015053e-01  1.00000000 1.000000000  0.09153177  0.44807046
#> Variable 2 4.806759e-05  0.62502692 1.000000000  0.01196223  0.01774364
#> Variable 3 1.000000e+00  0.18598474 0.045159915  1.00000000  1.00000000
#> Variable 4 1.000000e+00  0.02026738 0.001539579  1.00000000  1.00000000
#> Variable 5 1.000000e+00  1.00000000 1.000000000  1.00000000  1.00000000
#>             Variable 17  Variable 18 Variable 19 Variable 20
#> Variable 1 1.0000000000 1.000000e+00  1.00000000  1.00000000
#> Variable 2 0.0001415309 1.000000e+00  1.00000000  0.04101737
#> Variable 3 0.1009277467 9.609877e-03  1.00000000  0.46480405
#> Variable 4 1.0000000000 5.904616e-05  0.02731225  0.12432018
#> Variable 5 1.0000000000 9.602709e-01  0.55637338  1.00000000
#>